+++ priv/repo/migrations/20190327135106_create_transactions.exs
@@ -7,7 +7,7 @@ defmodule BankChallenge.Repo.Migrations.CreateTransactions do
       add :name, :string
       add :account_number, references(:accounts, column: :account_number, type: :uuid)
       add :from_account_number, references(:accounts, column: :account_number, type: :uuid), null: true
-      add :amount, :decimal
+      add :amount, :integer
       
       timestamps()
     end
